reorganise logic, refactor code
